refactor(dts): split declaration statement emit
